decomposition.py

+++ emgdecomp/decomposition.py
@@ -842,7 +842,7 @@ class EmgDecomposition(object):
 
     def _firings_dtype(self):
         return np.dtype([
-            ('source_idx', int),
+            ('source_idx', np.int),
             ('discharge_samples', np.int),
             ('discharge_seconds', np.float),
             ('squared_amplitude', np.float),
